Add -f flag to start in foreground and log to stdout
Running with the -d flag was the only way to have the process remain in the foreground. That can at the price of a lot of log messages. We've split the foregrounding from the -d into its own command line switch -f.

Additionally we've remove logging to a file. Instead log messages are written to stdout. This will make astmanproxy more container friendly.